+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added support for [using an Address in GraphQL](https://plugins.doublesecretagency.com/google-maps/address-field/graphql/).
 - Added support for [reverse proximity searches](https://plugins.doublesecretagency.com/google-maps/guides/reverse-proximity-search/).
 - Added support for [drawing circles on a map](https://plugins.doublesecretagency.com/google-maps/guides/drawing-circles/).
+- Added [`hideCircle`](https://plugins.doublesecretagency.com/google-maps/dynamic-maps/universal-methods/#hidecircle-circleid) universal method.
+- Added [`showCircle`](https://plugins.doublesecretagency.com/google-maps/dynamic-maps/universal-methods/#showcircle-circleid) universal method.
 - Added [`getCircle`](https://plugins.doublesecretagency.com/google-maps/dynamic-maps/javascript-methods/#getcircle-circleid) JavaScript method.
 
 ### Changed